Make Key.getKey return a new button each time
The special keys exist in each layer, and the code was previously
trying to add the same St.Button to each layer's group, which caused
warnings and made the actors disappear. Fix this by creating a new
St.Button actor each time we see one of the special key names.