Fix scoping for extended keys
when creating an anonymous function inside a loop, you need to make
sure that the variables used by the function are recreated in each
iteration of the loop.

"for each (extended_key in ..." refers to a global variable
"extended_key", which means that every button-press-event handler
ended up referring to the same "extended_key" variable, and so they
all ended up typing the same thing.

Changing that to "for each (let extended_key in ..." creates a new
extended_key variable for each call to _getExtendedKeys(), but still
reuses that variable for each iteration of the loop, so all of the
extended keys for each character end up typing the value of the last
extended key in that set.

Changing it to have "let extended_key" *inside* the loop fixes
everything.