keyboard: fix dragging
Normally the stage has holes everywhere there isn't a piece of chrome,
so if you drag too quickly and the pointer leaves the outline of the
keyboard, then it will leave the stage as well, and so clutter will
stop emitting motion events.

Fix this by setting the stage_input_mode to FULLSCREEN when dragging.